Update sources from github

+++ Engine/Graphics/YnText.cs
@@ -189,7 +189,7 @@ namespace Yna.Engine.Graphics
                 _assetLoaded = true;
 
                 UpdateTextMeasure();
-                Rectangle = new Rectangle(X, Y, (int)_tempMeasure.X, (int)_tempMeasure.Y);
+                Rectangle = new Rectangle((int)X, (int)Y, (int)_tempMeasure.X, (int)_tempMeasure.Y);
             }
         }
